+++ libraries/exoplayer_dash/src/test/java/androidx/media3/exoplayer/dash/DefaultDashChunkSourceTest.java
@@ -644,6 +644,128 @@ public class DefaultDashChunkSourceTest {
   }
 
   @Test
+  public void getNextChunk_imageTrackWithPresentationTimeOffset_setsCorrectSampleOffsetUs()
+      throws Exception {
+    String mpdXml =
+        "<MPD xmlns=\"urn:mpeg:dash:schema:mpd:2011\""
+            + " profiles=\"urn:mpeg:dash:profile:isoff-live:2011\" minBufferTime=\"PT1.5S\""
+            + " mediaPresentationDuration=\"PT10S\">\n"
+            + "  <Period id=\"0\" start=\"PT0S\">\n"
+            + "    <AdaptationSet contentType=\"image\" mimeType=\"image/jpeg\">\n"
+            + "      <SegmentTemplate timescale=\"1000\" presentationTimeOffset=\"50000000\""
+            + " media=\"https://example.com/thumb_$Time$.jpg\">\n"
+            + "        <SegmentTimeline>\n"
+            + "          <S t=\"50000000\" d=\"5000\" r=\"10\"/>\n"
+            + "        </SegmentTimeline>\n"
+            + "      </SegmentTemplate>\n"
+            + "      <Representation id=\"0\" bandwidth=\"10000\" width=\"192\" height=\"108\"/>\n"
+            + "    </AdaptationSet>\n"
+            + "  </Period>\n"
+            + "</MPD>";
+    DashManifest manifest =
+        new DashManifestParser()
+            .parse(
+                Uri.parse("https://example.com/test.mpd"),
+                new ByteArrayInputStream(mpdXml.getBytes(UTF_8)));
+    FakeDataSet fakeDataSet =
+        new FakeDataSet()
+            .newData("https://example.com/thumb_50000000.jpg")
+            .appendReadData(new byte[] {(byte) 0xFF, (byte) 0xD8, (byte) 0xFF, (byte) 0xE0})
+            .endData();
+    FakeDataSource fakeDataSource = new FakeDataSource(fakeDataSet);
+    Format trackFormat = manifest.getPeriod(0).adaptationSets.get(0).representations.get(0).format;
+    DefaultDashChunkSource chunkSource =
+        new DefaultDashChunkSource(
+            new BundledChunkExtractor.Factory(),
+            new LoaderErrorThrower.Placeholder(),
+            manifest,
+            new BaseUrlExclusionList(),
+            /* periodIndex= */ 0,
+            /* adaptationSetIndices= */ new int[] {0},
+            new FixedTrackSelection(new TrackGroup(trackFormat), /* track= */ 0),
+            C.TRACK_TYPE_IMAGE,
+            fakeDataSource,
+            /* elapsedRealtimeOffsetMs= */ 0,
+            /* maxSegmentsPerLoad= */ 1,
+            /* enableEventMessageTrack= */ false,
+            /* closedCaptionFormats= */ ImmutableList.of(),
+            /* playerTrackEmsgHandler= */ null,
+            PlayerId.UNSET,
+            /* cmcdConfiguration= */ null);
+    ChunkHolder output = new ChunkHolder();
+
+    chunkSource.getNextChunk(
+        new LoadingInfo.Builder().setPlaybackPositionUs(0).build(),
+        /* loadPositionUs= */ 0,
+        /* queue= */ ImmutableList.of(),
+        output);
+    BaseMediaChunk mediaChunk = (BaseMediaChunk) output.chunk;
+    SampleQueue sampleQueue =
+        SampleQueue.createWithoutDrm(
+            new DefaultAllocator(/* trimOnReset= */ true, C.DEFAULT_BUFFER_SEGMENT_SIZE));
+    BaseMediaChunkOutput chunkOutput =
+        new BaseMediaChunkOutput(new int[] {C.TRACK_TYPE_IMAGE}, new SampleQueue[] {sampleQueue});
+    mediaChunk.init(chunkOutput);
+    mediaChunk.load();
+
+    assertThat(sampleQueue.getLargestQueuedTimestampUs()).isEqualTo(0);
+  }
+
+  @Test
+  public void getNextChunk_correctlySetsLocationOnDataSpec() throws Exception {
+    String mpdXml =
+        "<MPD xmlns=\"urn:mpeg:dash:schema:mpd:2011\""
+            + " profiles=\"urn:mpeg:dash:profile:isoff-live:2011\" minBufferTime=\"PT1.5S\""
+            + " mediaPresentationDuration=\"PT10S\">\n"
+            + "  <BaseURL serviceLocation=\"cdn_east\">https://example.com/video/</BaseURL>\n"
+            + "  <Period id=\"0\" start=\"PT0S\">\n"
+            + "    <AdaptationSet id=\"0\" contentType=\"video\">\n"
+            + "      <Representation id=\"0\" bandwidth=\"10000\" width=\"192\" height=\"108\">\n"
+            + "        <SegmentList duration=\"1000\">\n"
+            + "          <Initialization sourceURL=\"init.mp4\"/>\n"
+            + "          <SegmentURL media=\"seg-1.mp4\"/>\n"
+            + "        </SegmentList>\n"
+            + "      </Representation>\n"
+            + "    </AdaptationSet>\n"
+            + "  </Period>\n"
+            + "</MPD>";
+    DashManifest manifest =
+        new DashManifestParser()
+            .parse(
+                Uri.parse("https://example.com/test.mpd"),
+                new ByteArrayInputStream(mpdXml.getBytes(UTF_8)));
+    Format trackFormat = manifest.getPeriod(0).adaptationSets.get(0).representations.get(0).format;
+    DefaultDashChunkSource chunkSource =
+        new DefaultDashChunkSource(
+            new BundledChunkExtractor.Factory(),
+            new LoaderErrorThrower.Placeholder(),
+            manifest,
+            new BaseUrlExclusionList(),
+            /* periodIndex= */ 0,
+            /* adaptationSetIndices= */ new int[] {0},
+            new FixedTrackSelection(new TrackGroup(trackFormat), /* track= */ 0),
+            C.TRACK_TYPE_VIDEO,
+            new FakeDataSource(),
+            /* elapsedRealtimeOffsetMs= */ 0,
+            /* maxSegmentsPerLoad= */ 1,
+            /* enableEventMessageTrack= */ false,
+            /* closedCaptionFormats= */ ImmutableList.of(),
+            /* playerTrackEmsgHandler= */ null,
+            PlayerId.UNSET,
+            /* cmcdConfiguration= */ null);
+    ChunkHolder output = new ChunkHolder();
+
+    chunkSource.getNextChunk(
+        new LoadingInfo.Builder().setPlaybackPositionUs(0).build(),
+        /* loadPositionUs= */ 0,
+        /* queue= */ ImmutableList.of(),
+        output);
+
+    assertThat(output.chunk).isNotNull();
+    assertThat(output.chunk.dataSpec.location).isEqualTo("cdn_east");
+  }
+
+  @Test
   public void updateManifest_representationWithZeroSegments_doesNotThrow() throws Exception {
     DashManifestParser parser = new DashManifestParser();
     DashManifest emptyManifest =
@@ -752,73 +874,5 @@ public class DefaultDashChunkSourceTest {
             new byte[0]);
     return new LoadErrorHandlingPolicy.LoadErrorInfo(
         loadEventInfo, mediaLoadData, invalidResponseCodeException, errorCount);
-  }
-
-  @Test
-  public void getNextChunk_imageTrackWithPresentationTimeOffset_setsCorrectSampleOffsetUs()
-      throws Exception {
-    String mpdXml =
-        "<MPD xmlns=\"urn:mpeg:dash:schema:mpd:2011\""
-            + " profiles=\"urn:mpeg:dash:profile:isoff-live:2011\" minBufferTime=\"PT1.5S\""
-            + " mediaPresentationDuration=\"PT10S\">\n"
-            + "  <Period id=\"0\" start=\"PT0S\">\n"
-            + "    <AdaptationSet contentType=\"image\" mimeType=\"image/jpeg\">\n"
-            + "      <SegmentTemplate timescale=\"1000\" presentationTimeOffset=\"50000000\""
-            + " media=\"https://example.com/thumb_$Time$.jpg\">\n"
-            + "        <SegmentTimeline>\n"
-            + "          <S t=\"50000000\" d=\"5000\" r=\"10\"/>\n"
-            + "        </SegmentTimeline>\n"
-            + "      </SegmentTemplate>\n"
-            + "      <Representation id=\"0\" bandwidth=\"10000\" width=\"192\" height=\"108\"/>\n"
-            + "    </AdaptationSet>\n"
-            + "  </Period>\n"
-            + "</MPD>";
-    DashManifest manifest =
-        new DashManifestParser()
-            .parse(
-                Uri.parse("https://example.com/test.mpd"),
-                new ByteArrayInputStream(mpdXml.getBytes(UTF_8)));
-    FakeDataSet fakeDataSet =
-        new FakeDataSet()
-            .newData("https://example.com/thumb_50000000.jpg")
-            .appendReadData(new byte[] {(byte) 0xFF, (byte) 0xD8, (byte) 0xFF, (byte) 0xE0})
-            .endData();
-    FakeDataSource fakeDataSource = new FakeDataSource(fakeDataSet);
-    Format trackFormat = manifest.getPeriod(0).adaptationSets.get(0).representations.get(0).format;
-    DefaultDashChunkSource chunkSource =
-        new DefaultDashChunkSource(
-            new BundledChunkExtractor.Factory(),
-            new LoaderErrorThrower.Placeholder(),
-            manifest,
-            new BaseUrlExclusionList(),
-            /* periodIndex= */ 0,
-            /* adaptationSetIndices= */ new int[] {0},
-            new FixedTrackSelection(new TrackGroup(trackFormat), /* track= */ 0),
-            C.TRACK_TYPE_IMAGE,
-            fakeDataSource,
-            /* elapsedRealtimeOffsetMs= */ 0,
-            /* maxSegmentsPerLoad= */ 1,
-            /* enableEventMessageTrack= */ false,
-            /* closedCaptionFormats= */ ImmutableList.of(),
-            /* playerTrackEmsgHandler= */ null,
-            PlayerId.UNSET,
-            /* cmcdConfiguration= */ null);
-    ChunkHolder output = new ChunkHolder();
-
-    chunkSource.getNextChunk(
-        new LoadingInfo.Builder().setPlaybackPositionUs(0).build(),
-        /* loadPositionUs= */ 0,
-        /* queue= */ ImmutableList.of(),
-        output);
-    BaseMediaChunk mediaChunk = (BaseMediaChunk) output.chunk;
-    SampleQueue sampleQueue =
-        SampleQueue.createWithoutDrm(
-            new DefaultAllocator(/* trimOnReset= */ true, C.DEFAULT_BUFFER_SEGMENT_SIZE));
-    BaseMediaChunkOutput chunkOutput =
-        new BaseMediaChunkOutput(new int[] {C.TRACK_TYPE_IMAGE}, new SampleQueue[] {sampleQueue});
-    mediaChunk.init(chunkOutput);
-    mediaChunk.load();
-
-    assertThat(sampleQueue.getLargestQueuedTimestampUs()).isEqualTo(0);
   }
 }